fix: setting a 7x larger batch size does not make it run 7x faster
i don't get it. For now, we're not going to look at it. We're just going to look away and pretend it's not happening. There must be some sort of bottleneck somewhere that is meaning that it is taking forever to do something when you have a larger batch size. This might need looking at later