test: add torch device fixture (with auto-skipping) for running on cuda
To be able to know that things are ready to go on any type of machine,
we can run all the tests on both CPU and GPU. however, if we don't have
access to the gpus locally, then it can causes errors but with the
awesomeness of pytest, we can skip if cuda is not available! woo!

There's also the possibility that a test might be skipped if cuda is not
available but we are expecting it to be. I'm not sure what to do in this
situation but I guess I'll cross the bridge when it happens?