feat(metrics): just use the same set of metrics for training and validation
after a week and a bit of debugging, i found that i used too many metrics for the validation, which meant that it was deadlocking during the validation epoch, which is just a big big problem, so we just dont anymore.

simpler metrics and the damn things works. Everyone wins.