feat!: change how assets are structured to get easily get instance metadata for statistics dumping
A bunch chonker of a change because I didn't forsee this coming. To get all the details I need out of an instance, I want to also get the objects used in the "frame" placeholder types in prompts so we can see distributions and the like.

So, to get everything, needed to change some of the structures, which mean the features needed to change, which means the raw datasets need to be patched. Thankfully, the previous patch _should_ work nicely.